+++ frontEnd/src/pages/create-issue-page/NewIssueBtnFooter.js
